% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap_create_metadata_structure.R
\name{create_dictionary}
\alias{create_dictionary}
\title{Create data dictionary}
\usage{
create_dictionary(df)
}
\arguments{
\item{parameter_list_raw}{list}
}
\value{
list
}
\description{
Create data dictionary
}
