% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_redcap_var_name}
\alias{create_redcap_var_name}
\title{Take a string and build redcap variable name}
\usage{
create_redcap_var_name(strg)
}
\arguments{
\item{strg}{a string}
}
\value{
a redcap variable name
}
\description{
Take a string (usage is that it is the text of a question in the paramter) and
clean name (remove major case and space), then take from 1 to 26 (to comply
with RCP rules in terms of variable naming)
}
