% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter_only_redcap_columns}
\alias{filter_only_redcap_columns}
\title{Filter only redcap column}
\usage{
filter_only_redcap_columns(dic)
}
\arguments{
\item{dic}{dataframe}
}
\value{
dataframe
}
\description{
Filter only redcap column
}
