% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_table_recoding_for_redcap}
\alias{get_table_recoding_for_redcap}
\title{Table to translate field type from param list to RCp dictionary}
\usage{
get_table_recoding_for_redcap()
}
\value{
a dataframes
}
\description{
Table with same column as a Redcap dictionary and a start to translate
'the information from the excel parameter list into data dictionary information
(field types and some form of validation for dates, integer and date + time)
}
