% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap_create_metadata_structure.R
\name{read_parameter_list}
\alias{read_parameter_list}
\title{Read parameter list}
\usage{
read_parameter_list(path_parameter_excel_workbook)
}
\arguments{
\item{path_parameter_excel_workbook}{path to parameter list excel workbook}
}
\value{
list
}
\description{
Read parameter list
}
