% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap_create_metadata_structure.R
\name{redcap_create_metadata_structure}
\alias{redcap_create_metadata_structure}
\title{Transform a parameter list into a REDCap project}
\usage{
redcap_create_metadata_structure(
  credential_path,
  pid,
  path_parameter_list,
  path_xlsx_output,
  project_name
)
}
\arguments{
\item{credential_path}{Path where the credential file is stored}

\item{pid}{Project ID}

\item{path_parameter_list}{path to parameter list excel workbook}

\item{path_xlsx_output}{path where draft of data dictionary is saved}

\item{project_name}{project name (is used for naming the xlsx file)}
}
\value{
list
}
\description{
This function takes a CTC-ZH parameter list as input and tranform it into
a REDCap project and an excel file that contain a. an excel sheet with the
corresponding data dictionary and b. an excel sheet with the corresponding
data dictionary + relevant annotations to finalize the REDCap project
}
