% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_accent}
\alias{remove_accent}
\title{Remove acccent from a string}
\usage{
remove_accent(txt)
}
\arguments{
\item{txt}{a string}
}
\value{
a string without any accent
}
\description{
Remove acccent from a string
}
