% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap_create_metadata_structure.R
\name{reshape_parameter_list}
\alias{reshape_parameter_list}
\title{Reshape parameter list}
\usage{
reshape_parameter_list(parameter_list_raw)
}
\arguments{
\item{parameter_list_raw}{list}
}
\value{
list
}
\description{
Reshape parameter list
}
