% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{search_index_str_in_df}
\alias{search_index_str_in_df}
\title{Search index string in data.frame}
\usage{
search_index_str_in_df(str, df)
}
\arguments{
\item{df}{a dataframe}

\item{string}{a string}
}
\value{
vector of index
}
\description{
Search index string in data.frame
}
