% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{split_and_count_choice_options}
\alias{split_and_count_choice_options}
\title{Count number of choice option}
\usage{
split_and_count_choice_options(string_choice)
}
\arguments{
\item{string_choice}{string choice form parameter list}
}
\value{
an integer
}
\description{
take choice option as formatted in the parameter and return number of choices
}
