% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{split_and_enumerate_choice_options}
\alias{split_and_enumerate_choice_options}
\title{Format choice options for redcap}
\usage{
split_and_enumerate_choice_options(string_choice)
}
\arguments{
\item{string_choice}{string choice form parameter list}
}
\value{
formatted choice options
}
\description{
take choice option as formatted in the parameter list and return a string with
choice options formatted for redcap
}
